requirements.txt
numpy
scipy
sounddevice
paho-mqtt
loguru
PyYAML
tflite-runtime
requests
